Update JNI libjvm example to build on Linux.
There are platform-specific headers for JNI. There isn't a Maven property we can use to refer to the OS family so we just list both here. A JVM or GraalVM distribution will only include the directory specific to the platform it was built for. The other directory will not be present and clang++ will happily ignore it.