Move constant definitions into the Truffle language Haversine distance implementations to avoid conflicts re-evaluating the code snippet within the same context.
We want to use one context for all benchmark runs, but the Google Benchmark runner will execute the benchmark function (and its setup) multiple times during an benchmark run. That ends up evaluating the snippet multiple times. Since we're using anonymous functions for the implementation, there are no conflicts with the distance methods, but trying to redefine a constant is porblematic.