Add benchmarks (#2)
* Added a new launcher for benchmarking.

The launcher can switch between multiple modes of calling a Haversine distance function with different languages. The idea is to eventually measure the call overhead of the various mechanisms for executing code in a Native Image-generated shared library.

* Reuse the context for `@CEntryPoint` calls using Truffle polyglot API.

Creating and destroying a context per-call is nice and clean, but it means each call starts from scratch so there's no opportunity to JIT anything. This code is now relying on Graal isolates to clean up the context on teardown and keeping the static values unique in each isolate.

* Add JNI benchmarks.

* Move constant definitions into the Truffle language Haversine distance implementations to avoid conflicts re-evaluating the code snippet within the same context.

We want to use one context for all benchmark runs, but the Google Benchmark runner will execute the benchmark function (and its setup) multiple times during an benchmark run. That ends up evaluating the snippet multiple times. Since we're using anonymous functions for the implementation, there are no conflicts with the distance methods, but trying to redefine a constant is porblematic.

* Add a build-time variable for controlling whether a Truffle context is reused between benchmark runs.

* Reformat all C/C++ code with `clang-format` for consistency.

* Rework benchmarks to take arguments rather than hard-code values.

* Ignore all target directories.

* Favor `nullptr` over `NULL` per `clang-tidy` recommendation.

* Apply some additional c++ best practices

* Prefer using c++ headers over c headers. * Always initialize pointers to nullptr. * Avoid using namespace std as it introduces ambiguity into function calls. * Ensure const-correctness with function parameters. * Prefer enum class over plain enum for better type safety. * Always initialize the first entry in an enum or enum class.

* Follow up fixes to a PR I merged prematurely.

* Dumped the `clang-format` Google style to a local file for better tool interop.

E.g., IDEs can read the .clang-format file to provide code formatting directly in the editor.

* Add a C++ implementation of the Haversine distance as a baseline benchmark.

* Pass arguments to configure common benchmark rather duplicate code.

* Add a non-thread-safe parse cache option for `@CEntryPoint` polyglot method.

This new benchmark is intended to show `@CEntryPoint` with arbitrary guest language code without the overhead of a concurrent parse cache.

* Compile the benchmark runner with optimizations enabled.

The benchmarks contain the C++ implementation of the Haversine function. Since Native Image is performing AOT optimizations, it's only fair to allow clang++ to do the same.

* Add more benchmarks for `@CEntryPoint` polyglot calls to get a better idea of the cost and benefit of a parse cache.

* Add another benchmark for `@CEntryPoint` and Truffle polyglot without context re-use.

* Clean up the benchmark code quite a bit.

Notably:
* Remove code related to original benchmark launcher
* Add options for dumping Graal graphs
* Move setup code to relevant benchmarks

* Ensure everyone setting up the Google Benchmark library is using the same version.

* Update the project to run with GraalVM 22.1.0.

* Upgrade jUnit from 4.11 to 4.13.2 to pick up a security fix.

* Update README for GraalVM 22.1.0.

Co-authored-by: Nabil Miri <nabil.miri@shopify.com>