Give each profile its own target directory.
Previously, each profile built to the same target directory and may conflict. It was simplest to just clear out the directory before starting a new build. But, that had the unfortunate side effect of discarding the build for any other profile. Some of these profiles take a while to build, so we don't want to toss good artifacts if we can avoid it.